feat:完成音乐上传功能,完善Button功能,累死老子了

+++ server/utils/writeIndex.js
@@ -1,8 +1,14 @@
 const fs = require('fs')
 const path = require('path')
+const debug = require('debug')('writeIndex')
 
 module.exports = function(){
-    const distIndexHtml = fs.readFileSync(path.resolve(__dirname,'../../public/static/index.html'))
-    fs.writeFileSync(path.resolve(__dirname,'../../public/index.html'),distIndexHtml)
-    console.log('写入index.html成功');
+    if(
+        fs.existsSync(path.resolve(__dirname,'../../public/static'))
+    ){
+        const distIndexHtml = fs.readFileSync(path.resolve(__dirname,'../../public/static/index.html'))
+        fs.writeFileSync(path.resolve(__dirname,'../../public/index.html'),distIndexHtml)
+        debug('写入index.html成功');
+    }
+
 }
